Common -> Document
'Common' was only used to keep the state of one document. Calling
it 'Document' makes it easier to think about the purpose of class.

When we add bulk document conversion, allowing us to keep track of
a document's state will be key. This name change is a step towards
that.