Remove window management logic
Since everyhting now happens in a single window, there is no need
to have a way to keep track other windows. They simply won't exist.

But on windows and linux it will still be possible to start
multiple windows by starting various dangerzone processes. On MacOS
this doesn't seem to be as easy from the launcher, but it should
not be critical as multiple documents can be converted at the same
time in the one window.