Hide widgets: select docs -> settings -> conversion
To help debugging and visualizing what was happening we set all
widgets to be visible at the same time. Now that is no longer needed so
we can hide them.

This keeps the original program flow:
  1. select the documents
  2. set the settings
  3. see the conversion progress

This diverges from the proposed design in issue #117 for simplification
and consistency (with past program flow) purposes.