FIXUP: try loading pyside2 first and then pyside6
The type checked mypy was producing countless error messages on linux
(where pyside2 is the default). Loading PySide2 before PySide6 seems to
resolve some of the issues.

The mypy --ignore-unused-imports had to be removed since it would fail
on systems with PySide6 and on systems with PySide2 they would complain
that the import of QtCore, for example, as being redefined.