Basic PoC Qubes support borrowing QubesPDFConverter
Ditches the Qubes Trusted PDF "Hack" and changes the Dangerzone client
to be compatible with the server component. Modifies server to support
the new protocol.

Protocol overview:
  - app qube -> disp qube:   send file
  - disp qube -> app qube:   send num_pages in two unsigned bytes
  - for each page:
    - disp qube -> app qube: send width  (2 unsigned bytes)
    - disp qube -> app qube: send height (2 unsigned bytes)
    - disp qube -> app qube: send pixel data (width * height * 3 bytes)

It also adds the PDF assembly part (which would happen in the 2nd
container). This is done through symlinking the container code. This is
of course a temporary measure.