Restructure container code & Qubes OS packaging
The files in `container/` no longer made sense to have that name since
the "document to pixels" part will run in Qubes OS in its own virtual
machine.

To adapt to this, this PR does the following:
  - moves all the files in `container` to `dangerzone/conversion`
  - splits the old `container/dangerzone.py` into its two components
    `dangerzone/conversion/{doc_to_pixels,pixels_to_pdf}.py` with a
    `common.py` file for shared functions
  - Moves the Dockerfile to the project root and adapts it to the new
    container code location
  - Adds an RPC file for dz.Convert (in `qubes/dz.Convert`)
  - Changes `setup.py` and `build-rpm.py` to be able to build to a Qubes
    target machine, with the command:

        ./install/linux/build-rpm.py --qubes

  - Adds instructions on how to build Dangerzone for Qubes

NOTE: following those instructions, running Dangerzone will fail with
this error:

  File "/usr/lib/python3.11/site-packages/dangerzone/conversion/doc_to_pixels.py", line 101, in convert
    mime = magic.Magic(mime=True)
           ^^^^^^^^^^^^^^^^^^^^^^
  TypeError: Magic.__init__() got an unexpected keyword argument 'mime'

This happens since the python-magic library that fedora packages is
outdated and Magic.__init__() still doesn't support the `mime` argument,
which we make use of.